requirements.txt
pillow
textwrap3

